requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
dnspython==2.4.2
elasticsearch==5.5.0
h11==0.14.0
html5tagger==1.3.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
motor==3.3.2
multidict==6.0.4
openai==1.6.1
packaging==23.2
pycryptodome==3.19.1
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
PyYAML==6.0.1
requests==2.31.0
sanic==23.12.0
Sanic-Cors==2.2.0
sanic-ext==23.12.0
sanic-routing==23.12.0
sanic-session==0.8.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
tracerite==1.1.1
typing_extensions==4.9.0
ujson==5.9.0
urllib3
uvloop==0.19.0
websockets==12.0
wheel==0.41.2
